Refactor timer view into viewmodel
